requirements.txt
# Requirements
azure-storage-blob
requests
tqdm
pyyaml